+++ cmd/hidden_lake/applications/remoter/internal/config/init.go
@@ -5,6 +5,7 @@ import (
 
 	hlr_settings "github.com/number571/go-peer/cmd/hidden_lake/applications/remoter/pkg/settings"
 	logger "github.com/number571/go-peer/internal/logger/std"
+	"github.com/number571/go-peer/pkg/crypto/random"
 )
 
 func InitConfig(cfgPath string, initCfg *SConfig) (IConfig, error) {
@@ -15,6 +16,7 @@ func InitConfig(cfgPath string, initCfg *SConfig) (IConfig, error) {
 		initCfg = &SConfig{
 			FSettings: &SConfigSettings{
 				FExecTimeoutMS: hlr_settings.CDefaultExecTimeout,
+				FPassword:      random.NewCSPRNG().GetString(32),
 			},
 			FLogging: []string{logger.CLogInfo, logger.CLogWarn, logger.CLogErro},
 			FAddress: &SAddress{