+++ cmd/hidden_lake/applications/remoter/pkg/settings/settings.go
@@ -6,7 +6,8 @@ const (
 )
 
 const (
-	CPathYML = "hlr.yml"
+	CPathYML        = "hlr.yml"
+	CHeaderPassword = "Hl-Remoter-Password" // nolint: gosec
 )
 
 const (