+++ examples/anonymity/remoter/_request/request.sh
@@ -7,6 +7,9 @@ PUSH_FORMAT='{
         "method":"POST",
 		"host":"hidden-lake-remoter",
 		"path":"/exec",
+        "head":{
+            "Hl-Remoter-Password": "DpxJFjAlrs4HOWga0wk14mZqQSBo9DxK"
+        },
         "body":"YmFzaFtAcmVtb3Rlci1zZXBhcmF0b3JdLWNbQHJlbW90ZXItc2VwYXJhdG9yXWVjaG8gJ2hlbGxvLCB3b3JsZCcgPiBmaWxlLnR4dCAmJiBjYXQgZmlsZS50eHQ="
     }
 }';