+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - Update `cmd/hidden_lake/applications/messenger`: delete passwords
 - Update `cmd/hidden_lake`: delete template service
+- Update `cmd/hidden_lake/service`: autodelete CHeaderResponseMode
 
 <!-- ... -->